add metrics logic

+++ backend/src/v2/mlflow/types/log.go
@@ -8,12 +8,12 @@ type LogParamRequest struct {
 }
 
 type LogMetricRequest struct {
-	RunId     string `json:"run_id"`
-	RunUUID   string `json:"run_uuid"`
-	Key       string `json:"key"`
-	Value     int64  `json:"value"`
-	Timestamp int64  `json:"timestamp"`
-	Step      int64  `json:"step"`
+	RunId     string  `json:"run_id"`
+	RunUUID   string  `json:"run_uuid"`
+	Key       string  `json:"key"`
+	Value     float64 `json:"value"`
+	Timestamp int64   `json:"timestamp"`
+	Step      int64   `json:"step"`
 }
 
 type LogInputRequest struct {
